add nonlinear systematics
The parameters for systematics can now be either a number (like before)
or a list of numbers, which become coefficients in a power series
expansion in the affected observable:

    delta = sum(delta_i * x^i)

Constraints can be specified for each coefficient, or omitted. The
entire systematic (the whole series) can also be fixed in the fit.

Also, "data" is now a list, and when sigex-ing an ensemble, this
sequence of signals defines what is the data for each fit. This is used
to feed in externally-produced fake data sets.